+++ domain/src/main/java/com/anytypeio/anytype/domain/base/Interactor.kt
@@ -1,0 +1,49 @@
+package com.anytypeio.anytype.domain.base
+
+import com.anytypeio.anytype.domain.base.Interactor.Status
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.withContext
+import kotlin.coroutines.CoroutineContext
+
+/**
+ * Interactor, whose invocation [Status] can be observed.
+ * It can be used as alternative for [BaseUseCase].
+ * @property context coroutine's context for this operation.
+ */
+abstract class Interactor<in P>(private val context: CoroutineContext = Dispatchers.IO) {
+
+    operator fun invoke(params: P): Flow<Status> {
+        return flow {
+            emit(Status.Started)
+            withContext(context) { run(params) }
+            emit(Status.Success)
+        }.catch { t ->
+            emit(Status.Error(t))
+        }
+    }
+
+    protected abstract suspend fun run(params: P)
+
+    /**
+     * Invocation status for operation.
+     */
+    sealed class Status {
+        /**
+         * Operation is running now.
+         */
+        object Started : Status()
+
+        /**
+         * Operation completed successfully.
+         */
+        object Success : Status()
+
+        /**
+         * Operation terminated with error, represented as a [Throwable].
+         */
+        data class Error(val throwable: Throwable) : Status()
+    }
+}